make open_project a class
