+++ packages/server/lib/gui/events.js
@@ -17,7 +17,7 @@ const user = require('../user')
 const errors = require('../errors')
 const Updater = require('../updater')
 const { ProjectBase } = require('../project-base')
-const openProject = require('../open_project')
+const { openProject } = require('../open_project')
 const ensureUrl = require('../util/ensure-url')
 const chromePolicyCheck = require('../util/chrome_policy_check')
 const browsers = require('../browsers')
