+++ packages/server/lib/modes/run-ct.js
@@ -1,3 +1,5 @@
+const { openProject } = require('../open_project')
+
 const run = (options) => {
   // TODO: make sure if we need to run this in electron by default to match e2e behavior?
   options.browser = options.browser || 'electron'
@@ -12,7 +14,7 @@ const run = (options) => {
     // if we get compilation error (e.g. from webpack) it means that no assets were emitted
     // and we can not run any tests, even if the error belongs to the different module
     // that's why the only way to avoid this is to close the process
-    require('../open_project').closeBrowser().then(() => {
+    openProject.closeBrowser().then(() => {
       process.exit(1)
     })
   })
