+++ packages/server/lib/modes/run.js
@@ -15,7 +15,7 @@ const errors = require('../errors')
 const { ProjectBase } = require('../project-base')
 const Reporter = require('../reporter')
 const browserUtils = require('../browsers')
-const openProject = require('../open_project')
+const { openProject } = require('../open_project')
 const videoCapture = require('../video_capture')
 const { fs } = require('../util/fs')
 const runEvents = require('../plugins/run_events')
