+++ packages/server/test/integration/cypress_spec.js
@@ -40,7 +40,7 @@ const browsers = require(`${root}lib/browsers`)
 const videoCapture = require(`${root}lib/video_capture`)
 const browserUtils = require(`${root}lib/browsers/utils`)
 const chromeBrowser = require(`${root}lib/browsers/chrome`)
-const openProject = require(`${root}lib/open_project`)
+const { openProject } = require(`${root}lib/open_project`)
 const env = require(`${root}lib/util/env`)
 const v = require(`${root}lib/util/validation`)
 const system = require(`${root}lib/util/system`)
