+++ packages/server/test/unit/gui/events_spec.js
@@ -13,7 +13,7 @@ const Updater = require(`${root}../lib/updater`)
 const user = require(`${root}../lib/user`)
 const errors = require(`${root}../lib/errors`)
 const browsers = require(`${root}../lib/browsers`)
-const openProject = require(`${root}../lib/open_project`)
+const { openProject } = require(`${root}../lib/open_project`)
 const open = require(`${root}../lib/util/open`)
 const auth = require(`${root}../lib/gui/auth`)
 const logs = require(`${root}../lib/gui/logs`)
@@ -800,27 +800,27 @@ describe('lib/gui/events', () => {
         return sinon.stub(ProjectBase.prototype, 'close').withArgs({ sync: true }).resolves()
       })
 
-      it('is noop and returns null when no project is open', function () {
-        expect(openProject.getProject()).to.be.null
+      it('is noop and returns undefined when no project is open', function () {
+        expect(openProject.getProject()).to.be.undefined
 
         return this.handleEvent('close:project').then((assert) => {
           return assert.sendCalledWith(null)
         })
       })
 
-      it('closes down open project and returns null', function () {
+      it('closes down open project and returns undefined', function () {
         sinon.stub(ProjectBase.prototype, 'getConfig').resolves({})
         sinon.stub(ProjectBase.prototype, 'open').resolves()
 
         return this.handleEvent('open:project', '/_test-output/path/to/project-e2e')
         .then(() => {
           // it should store the opened project
-          expect(openProject.getProject()).not.to.be.null
+          expect(openProject.getProject()).not.to.be.undefined
 
           return this.handleEvent('close:project')
         }).then((assert) => {
           // it should store the opened project
-          expect(openProject.getProject()).to.be.null
+          expect(openProject.getProject()).to.be.undefined
 
           return assert.sendCalledWith(null)
         })
