+++ packages/server/test/unit/gui/files_spec.ts
@@ -4,7 +4,7 @@ import { expect } from 'chai'
 import 'sinon-chai'
 
 import { showDialogAndCreateSpec } from '../../../lib/gui/files'
-import openProject from '../../../lib/open_project'
+import { openProject } from '../../../lib/open_project'
 import { ProjectBase } from '../../../lib/project-base'
 import * as dialog from '../../../lib/gui/dialog'
 import * as specWriter from '../../../lib/util/spec_writer'
