+++ packages/server/test/unit/modes/run_spec.js
@@ -14,7 +14,7 @@ const ProjectBase = require(`${root}../lib/project-base`).ProjectBase
 const browsers = require(`${root}../lib/browsers`)
 const Reporter = require(`${root}../lib/reporter`)
 const runMode = require(`${root}../lib/modes/run`)
-const openProject = require(`${root}../lib/open_project`)
+const { openProject } = require(`${root}../lib/open_project`)
 const videoCapture = require(`${root}../lib/video_capture`)
 const env = require(`${root}../lib/util/env`)
 const random = require(`${root}../lib/util/random`)
